+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Sample selection respects the current image or video filters.
 - Added sort-by support via GUI.
+- Added similarity selection option.
+- Show confusion matrix in evaluation results for object detection.
 
 ### Changed